+++ ChangeLog.md
@@ -1,5 +1,9 @@
-# CHANGELOG EASYCOMMISSION FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
+# Change Log
+All notable changes to this project will be documented in this file.
 
-## 1.0
+## Unreleased
 
-Initial version
+## Version 1.0
+
+- FIX : Add conf to choose if we want to filter on billing date instead of payment date by default *05/05/2021* - 1.0.1
+- FIX : Add invoice status to take into consideration only paid invoices *04/05/2021* - 1.0.1